Rename session-step.tsx to session-step-builder.tsx.
